Remove empty tab groups
This CL automatically removes tab
groups without any tabs under them.

As those are not useful to the
user, this wasn't set as an option.

Note that this CL intentionally doesn't
filter them when loading the tabGroups.
This is because a linear search when
starting the extension would penalize
users with a lot of saved tabs. They
will catch up over time due to the new
behavior (and manually should they wish).

TEST=Manually checked that it works when removing a tab, restoring one or all tabs